refactor : CaldendarUIView View, Model 리펙토링
